moved the scrolling page to a separate component
