+++ package.json
@@ -11,7 +11,6 @@
   },
   "dependencies": {
     "bootstrap": "^5.3.2",
-    "firebase": "^10.4.0",
     "react": "^18.2.0",
     "react-bootstrap": "^2.9.0",
     "react-dom": "^18.2.0"
